requirements.txt
Django==3.1.2
gunicorn==19.9.0
PyMySQL==0.10.1
django-debug-toolbar==2.2
django-extensions==3.0.5
Pillow==7.1.2
mysqlclient==1.4.0
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.6
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
django-appconf==1.0.4
django-betterforms==1.2
django-bootstrap3==14.1.0
django-braces==1.14.0
django-mediumeditor==1.0.0
django-multiupload==0.5.2
django-storages==1.10.1
django-widget-tweaks==1.4.8
entrypoints==0.3
idna==2.10
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
nose==1.3.7
notebook==6.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.13.0
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.24.0
rope==0.17.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.11
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
